added the small small files of the helper.ts
Add initial project files and setup.
